Add support for obfuscation (FF-2343)
